Refactor ExpenseBase to separate base package for clarity
Same reason as previous two commits
