+++ RetireSimple.Engine/Data/Expense/OneTimeExpense.cs
@@ -1,8 +1,10 @@
-﻿namespace RetireSimple.Engine.Data.Expense {
+﻿using RetireSimple.Engine.Data.Base;
+
+namespace RetireSimple.Engine.Data.Expense {
 	/// <summary>
 	/// Represents an Expense on an investment that occurs once
 	/// </summary>
-	public class OneTimeExpense : ExpenseBase {
+	public class OneTimeExpense : Base.Expense {
 		/// <summary>
 		/// The date the expense is applied to the investment
 		/// </summary>
